// src/api/models/DurationStats.ts
/* tslint:disable */
/* eslint-disable */
/**
 * radarvan
 * Stats for generals
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * Order statistics for one set of game lengths, all in minutes.
 * @export
 * @interface DurationStats
 */
export interface DurationStats {
    /**
     * 
     * @type {number}
     * @memberof DurationStats
     */
    count: number;
    /**
     * 
     * @type {number}
     * @memberof DurationStats
     */
    totalMinutes: number;
    /**
     * 
     * @type {number}
     * @memberof DurationStats
     */
    meanMinutes?: number | null;
    /**
     * 
     * @type {number}
     * @memberof DurationStats
     */
    medianMinutes?: number | null;
    /**
     * 
     * @type {number}
     * @memberof DurationStats
     */
    p10Minutes?: number | null;
    /**
     * 
     * @type {number}
     * @memberof DurationStats
     */
    p90Minutes?: number | null;
    /**
     * 
     * @type {number}
     * @memberof DurationStats
     */
    shortestMinutes?: number | null;
    /**
     * 
     * @type {number}
     * @memberof DurationStats
     */
    longestMinutes?: number | null;
}

/**
 * Check if a given object implements the DurationStats interface.
 */
export function instanceOfDurationStats(value: object): value is DurationStats {
    if (!('count' in value) || value['count'] === undefined) return false;
    if (!('totalMinutes' in value) || value['totalMinutes'] === undefined) return false;
    return true;
}

export function DurationStatsFromJSON(json: any): DurationStats {
    return DurationStatsFromJSONTyped(json, false);
}

export function DurationStatsFromJSONTyped(json: any, ignoreDiscriminator: boolean): DurationStats {
    if (json == null) {
        return json;
    }
    return {
        
        'count': json['count'],
        'totalMinutes': json['totalMinutes'],
        'meanMinutes': json['meanMinutes'] == null ? undefined : json['meanMinutes'],
        'medianMinutes': json['medianMinutes'] == null ? undefined : json['medianMinutes'],
        'p10Minutes': json['p10Minutes'] == null ? undefined : json['p10Minutes'],
        'p90Minutes': json['p90Minutes'] == null ? undefined : json['p90Minutes'],
        'shortestMinutes': json['shortestMinutes'] == null ? undefined : json['shortestMinutes'],
        'longestMinutes': json['longestMinutes'] == null ? undefined : json['longestMinutes'],
    };
}

export function DurationStatsToJSON(json: any): DurationStats {
    return DurationStatsToJSONTyped(json, false);
}

export function DurationStatsToJSONTyped(value?: DurationStats | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'count': value['count'],
        'totalMinutes': value['totalMinutes'],
        'meanMinutes': value['meanMinutes'],
        'medianMinutes': value['medianMinutes'],
        'p10Minutes': value['p10Minutes'],
        'p90Minutes': value['p90Minutes'],
        'shortestMinutes': value['shortestMinutes'],
        'longestMinutes': value['longestMinutes'],
    };
}